requirements.txt
# absl-py==2.2.2
# annotated-types==0.7.0
# anyio==4.9.0
# astunparse==1.6.3
# blinker==1.9.0
# certifi==2025.4.26
# charset-normalizer==3.4.2
# click==8.2.0
# colorama==0.4.6
# fastapi==0.115.12
# Flask==3.1.1
# flatbuffers==25.2.10
# gast==0.6.0
# google-pasta==0.2.0
# grpcio==1.71.0
# h11==0.16.0
# h5py==3.13.0
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.6
# keras==3.9.2
# libclang==18.1.1
# Markdown==3.8
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# ml_dtypes==0.5.1
# namex==0.0.9
# numpy==2.1.3
# opt-einsum==3.4.0
# optree==0.15.0
# packaging==25.0
# pillow==11.2.1
# pip==24.0
# protobuf==5.29.4
# pydantic==2.11.4
# pydantic_core==2.33.2
# Pygments==2.19.1
# python-multipart==0.0.20
# requests==2.32.3
# rich==14.0.0
# setuptools==65.5.0
# six==1.17.0
# sniffio==1.3.1
# starlette==0.46.2
# tensorboard==2.19.0
# tensorboard-data-server==0.7.2
# tensorflow==2.19.0
# tensorflow-io-gcs-filesystem==0.31.0
# termcolor==3.1.0
# typing_extensions==4.13.2
# typing-inspection==0.4.0
# urllib3==2.4.0
# uvicorn==0.34.2
# Werkzeug==3.1.3
# wheel==0.45.1
# wrapt==1.17.2

absl-py==2.2.2
anyio==4.9.0
astunparse==1.6.3
fastapi==0.115.12
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
h11==0.16.0
h5py==3.13.0
Jinja2==3.1.6
keras==3.9.2
libclang==18.1.1
MarkupSafe==3.0.2
ml_dtypes==0.5.1
numpy==2.1.3
opt-einsum==3.4.0
optree==0.15.0
pillow==11.2.1
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
typing_extensions==4.13.2
uvicorn==0.34.2
wrapt==1.17.2